+++ client/src/api/foodsharepoints.js
@@ -1,0 +1,5 @@
+import { get } from './base'
+
+export async function listFoodSharePoints (regionId) {
+  return await get(`/regions/${regionId}/foodSharePoints`)
+}